+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `as->` to the "Resolve macro as" list
 - New code actions: refactor cond to if, and refactor if to cond
 - Allow finding references of multiple definitions at cursor. #2176
+- Fix CLI format `:style/indent` support via project type flag.
 - Lots of bumps:
   - clj-kondo: 2025.12.24-20251229.120141-5
   - rewrite-clj: 1.2.50 -> 1.2.51